Fixing modifier ranges not updating on the right frame.
This animation wasn't ending up on the right frame when quickly moving the cursor left->right. This was an error with update cycle order of the text modifier group vs the text object itself. The modifier group needs to always update before the text object!

@JcToon showed this issue to me in the office today...

https://github.com/rive-app/rive/assets/454182/76692e1d-e670-4324-a8ac-8c8ce0748e52

Diffs=
b17671130 Fixing modifier ranges not updating on the right frame. (#5623)