feature: Adding high-level API for querying a text run from an artboard
Adding a high-level API for getting/setting text run values. Doing this instead of giving ownership over the TextValueRun object as it should be safer - we'll always do a check to make sure the Artboard is still up before grabbing a text run and getting/setting the text value. This is in case users clean up the instances and accidentally still try to access a text run.

Diffs=
b151faa41 feature: Adding high-level API for querying a text run from an artboard (#5670)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>