patch: provide a backup source to load WASM from in case unpkg goes down
In case the main CDN for where we pull the WASM file from goes down (it's happened a few times), we should provide a backup CDN to pull from. Folks have compared [jsdelivr](https://www.jsdelivr.com/) as another reliable source to unpkg, which also serves brotli-compressed assets.

Hard to write a test for this, as the tests use the WASM-in-JS build, and so it never really touches this `wasmURL` logic (and honestly hard to tell how to invoke a catch block because its all emscripten build stuff), but verified locally in the example app it's what we should see.

<img width="639" alt="Screenshot 2023-08-09 at 3 32 11 PM" src="https://github.com/rive-app/rive/assets/5685058/08a115ec-b171-4eca-bafe-7cad004f817a">

Diffs=
04b50db5c patch: provide a backup source to load WASM from in case unpkg goes down (#5776)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>