Fix issue with timeline events on first frame
Events that fire on the first frame (with a work area too) were not reporting. This was because the way we look for occurred events wouldn't catch an event that started at the same time as the playhead as presumably that would've been caught on the previous frame. This falls apart when the animation starts on the same frame as the one where an event (or multiple) are triggered.

Adds a fix by detecting this condition and a bunch of tests in both C++ and Dart for it.

Also revs the lightning bolt animation in the editor for triggering the event as this showed an issue with state contention in its state machine.

Diffs=
382a48cf8 Fix issue with timeline events on first frame (#6006)