patch: add URL sanitization on open URL events
As discussed, we'll want to cover some basic sanitization of URLs from open url events.

Found this library that seems well used - https://github.com/braintree/sanitize-url/tree/main

Can include it as a dependency, but it's also pretty small and straightforward as is. Can also add more tests if we decide this is a route we want to go down, otherwise, happy for other suggestions.

Diffs=
6927d699f patch: add URL sanitization on open URL events (#5996)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>