Add a canvas-lite and canvas-advanced-lite package for the web runtime
Creating a few "lite" variants of the canvas runtime to publish in our workflow:
- `@rive-app/canvas-lite` - high-level JS api
- `@rive-app/canvas-advanced-lite` - low-level WASM

The `lite` versions of the web runtime will not support Rive Text for now. It will also be ~3x smaller than it is today (relevant for those who want a small footprint on their app and want to use Rive for small animations)

Because the example apps in this runtime test against local `*-single` builds (which compiles WASM in the JS so no need to load in WASM yourself), also building `*-single` variants of these lite packages, BUT we probably don't care to publish these, so just building them in gitignored `build/` folders that the example apps can reference when wanting to test the "lite" runtime variants.

TODO:
- [x] Test Text APIs to make sure we can still call them and it won't break rendering

Diffs=
014af5fff Add a canvas-lite and canvas-advanced-lite package for the web runtime (#6079)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>