patch: Only clean up the Renderer object if it is a CanvasRenderer, as WebGL Renderer does not need to be deleted
Original issue: https://github.com/rive-app/rive-wasm/issues/333

Only the canvas renderer can have its emscripten object deleted on cleanup, but not the WebGL one. So just adding a quick check to only delete if it detects the runtime has a canvas renderer.

Diffs=
c579540d1 patch: Only clean up the Renderer object if it is a CanvasRenderer, as WebGL Renderer does not need to be deleted (#6037)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>