feat: only use webgl instead of webgl2 in iOS browsers to try and prevent crashing on iOS Safari
Going to try and experiment with using only `webgl` on iOS Safari (as opposed to webgl2) to try and prevent crashing issues with the offscreen context for coordinating meshes.

See: https://2dimensions.slack.com/archives/CLLCU09T6/p1703108707108469

Initially thought about just surfacing a high-level API flag users can set to use webgl2 or not, but I don't want to introduce something like that until we can really verify on a number of devices that this solution to use WebGL works alright for us.

Also revoking an object URL when we create it for the CanvasRenderImage image src, as suggested by web docs.

Diffs=
fc8896bc6 feat: only use webgl instead of webgl2 in iOS browsers to try and prevent crashing on iOS Safari (#6383)

Co-authored-by: Zachary Plata <plata.zach@gmail.com>