poc for adding "flush"
poc for adding flush.

I think we just need to add `flushCanvasRenderers` as `flush` (also add this for the flushSkiaRenderer) and that should address https://github.com/rive-app/rive-wasm/issues/336.

obviously needs clean up, but just wanna make sure its worth sticking this in.

is there a good way to get back to the "issue" request to enquire if this will sort the problem? or do we just update the api and check in?

Diffs=
ebff4893c poc for adding "flush" (#6237)

Co-authored-by: Maxwell Talbot <talbot.maxwell@gmail.com>
Co-authored-by: Zachary Plata <plata.zach@gmail.com>