Updating harfbuzz to 8.3.0
[This thread](https://2dimensions.slack.com/archives/C067PQQL3EK/p1708104838351439) had me looking through the harfbuzz codebase to see if we could change the way some of the symbols were exported (our renames was getting clashing defines).

That made me take a closer look at all the changes since 6.0.0 (the version we were pegged at). There are a ton of performance and stability changes since 6.0.0 that we really want (take a look at the [release page](https://github.com/harfbuzz/harfbuzz/releases) for a list).

So I took this opportunity to update our renames script, pin us to a new harfbuzz and branch it with some changes to disable exporting certain symbols that are normally accessed via defines in C++ but are exported as a courtesy to wrapping language bindings. We don't need those so I added a way to compile them out for us.

Diffs=
9d605a1fe Updating harfbuzz to 8.3.0 (#6652)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>