fix: synthetic browser events
After a touchend event, some browsers may fire synthetic mouse events (mouseover, mousedown, mousemove, mouseup) if the touch interaction did not cause any default action (such as scrolling). This is done to simulate the behavior of a mouse for applications that do not support touch events.

This is a problem because it results in the JS code reporting duplicated pointer events to Rive:
 - Confirmed fix Chrome desktop (device simulation)
 - Confirmed fix on iOS Safari (real device)

This PR introduces a solution to keep track of the previous event and to not send the synthetic mouse events if the
touch event was a click (touchstart -> touchend).

This is only needed when `isTouchScrollEnabled` is false. When true, `preventDefault()` is called which prevents this behaviour.

Not helpful for the solution, but for context:
- https://stackoverflow.com/questions/9656990/how-to-prevent-simulated-mouse-events-in-mobile-browsers
- https://stackoverflow.com/questions/25572070/javascript-touchend-versus-click-dilemma

Reported on Jira: https://rive.atlassian.net/browse/RIV-4845

Fixed solution (with logs):
https://github.com/user-attachments/assets/83a77c64-24b4-4381-8e97-178dfbcabc1d

Diffs=
bc9aab464 fix: synthetic browser events (#7609)

Co-authored-by: Gordon <pggordonhayes@gmail.com>