add viewmodel and data enum system core objects
This PR only has the boilerplate code to support system data enums.
System enums will be enums that are built-in in Rive, like text alignment, word wrapping, blend modes.
Those system enums won't be included in the .riv file, but they'll exist built-in in the runtime.
This PR creates 4 new core objects types. 2 extends from data_enum and 2 from viewmodel_property_enum.
The rest of the changes are renaming the base class to the custom one.

Diffs=
5297cf0cc2 add viewmodel and data enum system core objects (#8382)