invert order of advance between parent and child
we had inverted the order of advances between parent and child to fix a frame delay.
But this introduced a bug where the child would report an event, than in turn would trigger an event in the parent.
But immediately after, the parent would advance causing the event to get cleared.

Diffs=
01bc166a8e invert order of advance between parent and child (#8547)