Re-upload webgl texture data if context changes.
The editor disposes the canvas when going back to the file browser. This causes a new canvas to be created with a new WebGL context. We were caching images and their gl texture handles, which could belong to a previously created context. We now re-upload the texture data if we see the webgl context has changed.

This fixes the issue where the wrong image would be used for the grid pattern behind a transparent artboard.

Diffs=
7777300368 Re-upload webgl texture data if context changes. (#8655)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>