skip object on animation reset if it is null
after a recent change where we validate objects when we create an animation, if those objects are not valid, they get removed.
But animation_reset_factory was assuming they would never be nullptrs.
This checks whether the object exists, similar to how keyed_objects operate when applying an animation.

Diffs=
3c0cc1f329 skip object on animation reset if it is null (#8670)