Nnnn system data enums part 3
Adds system enums to list of available bindable enums.

There is no work needed on the cpp side, because we're mapping the values to integers and they translate directly to how our runtime consumes them.
In the future we might want to be more restrictive, but for now the way it works is very flexible.
Added comments in the PR

https://github.com/user-attachments/assets/9941fc9b-ac58-43dc-bcf6-9ebbd4ac4b55

Diffs=
c7c1701511 Nnnn system data enums part 3 (#8635)