Security warnings
trying to swat most of these https://github.com/rive-app/rive/security/dependabot?page=13&q=is%3Aopen+sort%3Amanifest-path+ecosystem%3Anpm

there are a few sets of issues here

`release-it` - i just updated these, we should be aware that on next release of android/ios/etc we may have an issue here, i did a dry run on rive_flutter which was fine mind you.

`_framework/examples` - i ran all of them successfully, *except for* the `vue` example which complains that we dont have `rive.lean.dev` in our dist. I actually suspect that this example stopped working but was hoping to confirm this @HayesGordon

`wasm/examples` - ran all of these successfully *except for* the `centaur_game` example. the scaling seems off, and the centaur is not in the frame properly, also the apples are being placed oddly & rendered "twice" once with a static apple and once with an animated one... I did a little playing around here, (changing a state machine to advanceAndApply etc) but once again was hoping to find out if this was actually working properly before. (will go and check out main and rebuild everything in a second to double check the state of this, but I couldnt even build this before updating packages... )

`runtime_wasm/js` hm this one is all just made up of dev bits, do we need to do more to test this than try the examples?

Diffs=
5893a12377 Security warnings (#8825)

Co-authored-by: Maxwell Talbot <talbot.maxwell@gmail.com>