resize canvas when visible
When the resize observer callback is triggered, we need to recalculate the canvas size.
This solves mutiple scenarios, but in particular when an element is set to display none, its bounding client rect will be zero, so the canvas will have no size. When the canvas display changes, we need to measure the canvas back to get its new size.

Diffs=
b9e34ece24 resize canvas when visible (#8915)

Co-authored-by: hernan <hernan@rive.app>