change inherit blend mode default value to 127
when -1 was serialized, it was being converted to 127 and would break the editor blend mode mapping.
so this PR changes the default value to 127.
Fortunately, older files that didn't change the blend mode explicitly wouldn't have any value, so this will still be the default.
And files that did change the blend value to something different than inherit, and changed it back to inherit, will have 127 as value that will match the new inherit assigned value.
So this covers older files and newer files with old runtimes and new runtimes.

Diffs=
bf57f85d7f change inherit blend mode default value to 127 (#8926)