Change Scroll DraggableProxy to use drawProxy
Currently when the state machine processes DraggableConstraints, it uses the viewport layout as it's "hitLayout". After the hit objects are sorted, this means the hitLayout is processed first, and can prevent listeners attached to other drawables inside the scroll viewport from firing when the pointer move event is being processed.

@LauraRive came across a problem when trying to add a draggable slider inside the scrolling area. In this case, the slider does not respond because the DraggableConstraint's hitLayout is processed first. This PR changes the constraint to use the layout's proxy as its hittable, which processes it after it's contents, meaning other listeners within the scroll area can be processed first. This change also properly enables having nested scrollable areas (for example horizontal scroll rows inside a vertically scrolling area.

Diffs=
834ec98f33 Change Scroll DraggableProxy to use drawProxy (#9084)