Rename gold/ to zzzgold/
GitHub only looks at the first ~300 changed files in a PR when determining which tests to run. Give the gold directory a name that starts with z so it doesn't overwhelm GitHub's workflow decision process with nothing but updated images. It's also nicer to have the image diffs last in a PR.

Diffs=
16567a3417 Rename gold/ to zzzgold/ (#9108)