library: view model support (#9630) 16c30e956a This PR allows drag+drop a ViewModel from a library asset into a host file. I tried to keep it readable via commits.
The approach is pretty similar to how DataEnums work with libraries. I've wrapped every view model property into a library version, which would also contain a componentId that contains the ID of that property in the original library file's core.

I've added LibraryArtboard.viewModelId that's similar to Artboard.viewModelId. This is inferred from interalCore when the library file loads (in LibraryAsset.decode, `_updateVms`).

The runtime exporter's context needs to keep track of the instances we want to export from the host file. This is b/c view model instances have to be exported with view models, so the host's instances need to be passed back to the library file and exported with its view models. (mapping_context, `requiredVmInstances`).

Because runtime is extremely order sensitive, I've also added support for DataEnumValue.order and ViewModelProperty.propertyIndex in this PR.

QoL improvements: I refactored the caching logic to account for Library components that are removed, and it's now more generalized (this is tested).

Testing: I also added to the import export test in the editor and C++

Demo: https://2dimensions.slack.com/archives/C07M7DQL4F2/p1746745186933889

Fixes: https://github.com/rive-app/rive/issues/9461
Fixes: https://github.com/rive-app/rive/issues/9345
Fixes: https://github.com/rive-app/rive/issues/9570
Fixes: https://github.com/rive-app/rive/issues/9588