Use markHostTransformDirty instead of markHostingLayoutDirty when appropriate (#10632) af3f673d77 Address a couple of issues:
When an artboard has a host, at the time the artboard has a layout marked dirty, we mark the hosting layout dirty (if it is a LayoutNodeProvider such as a NestedArtboardLayout or ArtboardComponentList). In cases where the host is not a LayoutNodeProvider (ie, NestedArtboardLeaf) we should also mark the host's transform dirty so it can redraw itself using the Artboard's updated size. This surfaced recently when we started using the artboard's hugged size with NestedArtboardLeaf mode.
In Artboard xChanged/yChanged we should mark the host's transform dirty rather than marking layout dirty (because we don't want to recompute layout at that time).